Switch to WAVE audio recording for now
I couldn't get opuslib to work correctly. After encoding, there would
only be a few bytes of data. Also, writing an opus stream to a file
would require building the file header manually because that is not
supported by opuslib.

We might want to revisit this in the future. Opus is a uch better
codec for us. Using this method, we cannot compress the audio, which
means that audio files (especially long recordings) can be very large.
Therefore, the upload takes a long time and a lot of space is occupied
by the file, which is potentially a bad thing (dropbox storage space is
very limited for free users, for example.)