Implement 'instantshare audio' command
Work in progress. Currently, there is a lot of code just copied over
from screen.py. As stated in the code comments starting in line 30,
we should move this code to a central location, so it can be reused.

The resulting URL from uploading to dropbox now offers you to download
the file, which is not very compelling to the people the link is shared
with, I feel. Maybe we can do something about this.