Basic functionality
"instantshare text" reads text from the clipboard, "instantshare text
--stdin" reads everything passed to its stdin. In interactive mode, only
a single line is read this way, but text files can be passed like this:

  instantshare text --stdin </path/to/file

This does not work from IDEs like pycharm though, the expression is not
evaluated by the shell.

The data is then saved to a temporary file and gets uploaded to the
selected hoster. This step is neccessary because that's the way our
interface to the storage providers works. We could think about changing
this in the future.